requirements.txt
torch==2.7.1
torchvision==0.22.1
opencv-python==4.12.0.88
numpy==2.2.6
supervision==0.27.0
inference-gpu
transformers
num2words
tqdm==4.67.1
python-dotenv==1.1.1
pyyaml==6.0.2
gdown==5.2.0
matplotlib==3.10.3
git+https://github.com/roboflow/sports.git@feat/basketball
